api/requirements.txt
tensorflow==2.12.0
flask==2.3.2
flask-cors==4.0.0
pillow==9.5.0
numpy>=1.22,<1.24
gunicorn==20.1.0